AI: Component agent ("Edit with AI") UX parity with the Assistant (#40)
Three parity items on the LIVE Component-agent surface (AIBuilderPage — the
page reached via /design/components/ai/:id; AIBuilderModal is dead/unrouted and
was left untouched):

1. Assistant-style tool-call rows. Extracted AssistantToolCallCard into a
   shared component (shared/AgentToolCallCard, class .agent-toolcall) used by
   BOTH surfaces — collapsible row with tool name, one-line arg summary, output
   size, and pretty-printed args/result via TerminalResponseBody. The Component
   agent previously rendered tool calls as bare blue Tags (name only).
   AssistantToolCallCard is now a back-compat re-export; the duplicated
   .assistant-toolcall SCSS was removed in favor of the shared .agent-toolcall.

2. Session export with secret-masking FROM THE START. New shared exporter
   (shared/exportAgentConversation) with maskSecrets() that redacts
   credential-named keys (password/api_key/token/secret/authorization/…)
   recursively from tool-call args+results before Markdown/JSON download. Wired
   into the Component agent's header via an export OverflowMenu. The Assistant's
   exportConversation is now a thin wrapper over the shared module, so it ALSO
   gets masking (it had none) for free — no forked copy.

3. Panel header renamed to "Edit with AI" (was "Edit Component with AI"); the
   new-component flow keeps its type-specific "Create {Chart/Control/Display}
   with AI".

Built on shared modules per the converge-on-shared-functions direction; both
agents now use one tool-call row + one exporter. Build 1998.
